requirements.txt
asgiref==3.7.2
asttokens==2.4.1
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.0
contourpy==1.1.1
cssselect2==0.7.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.7
djangorestframework==3.14.0
executing==2.0.1
filelock==3.13.1
fonttools==4.49.0
fsspec==2023.10.0
h5py==3.10.0
html5lib==1.1
huggingface-hub==0.21.4
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.3
ipykernel==6.26.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter-client==8.6.0
jupyter-core==5.5.0
kiwisolver==1.4.5
langdetect==1.0.9
MarkupSafe==2.1.3
matplotlib==3.7.5
matplotlib-inline==0.1.6
mplcursors==0.5.3
mpmath==1.3.0
nest-asyncio==1.5.8
networkx==3.1
nltk==3.8.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
packaging==23.2
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pgvector==0.2.3
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==4.0.0
prompt-toolkit==3.0.40
psutil==5.9.6
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyclipper==1.3.0.post5
pycparser==2.21
pydyf==0.9.0
Pygments==2.16.1
pyparsing==3.1.2
pypdf2==3.0.1
pypdfium2==4.27.0
pyphen==0.14.0
python-dateutil==2.8.2
python-doctr==0.8.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
rapidfuzz==3.6.2
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
shapely==2.0.3
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
sympy==1.12
threadpoolctl==3.2.0
tiktoken==0.5.1
tinycss2==1.2.1
tokenizers==0.14.1
torch==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
transformers==4.35.0
triton==2.1.0
typing-extensions==4.8.0
tzdata==2023.3
Unidecode==1.3.8
urllib3==2.0.7
wcwidth==0.2.9
weasyprint==61.1
webencodings==0.5.1
zipp==3.17.0
